feat: NJF contributions demo — SafetyNet capability + real blueprint (phase 5/6)
The full worked example for docs/guides/bulk-data-review.md: a new
validate-contributions-file SafetyNet Underwriting capability (automatic
deterministic rules in ContributionsValidation.cs — duplicate/missing member ref,
unrecognised tier, fire-endorsement surcharge floor, under-18/DOB consistency,
contribution-band warnings, deterministic member-ID matching — genuinely separate
from the existing human-decided risk-assessment queue, since this is automatic
underwriting with no staff involved), and njf-contributions.json wiring it all
together: upload -> materialize+call -> ingest -> review -> loop-back-to-resubmit
-> accept, gated by a real ServiceBlueprintRouteDefinition.ShowWhen
("contributionsErrorCount = 0").

Three real bugs found and fixed at the source while making this blueprint actually
run, none of them previously exercised because no existing blueprint used a
support-system/bulk-dataset-ingest output field inside a showWhen expression
before this one:
- CALC_FIELD_SHADOWS_INPUT false-positived on a genuine service-sourced
  calculations field, because Phase 1's own DataDisplay-known-field broadening had
  been folded into the same set that check used for "is this a captured input" —
  split into a narrower capturedInputFieldKeys for the shadow-check specifically.
- "source: service" calculation fields are resolved via a host-supplied
  serviceInputsResolver callback, not automatically from FieldValues — the
  reference app never wired one up (nothing needed it before). Added a generic
  one: reads whatever's declared source:"service" straight off the instance's own
  FieldValues, since the "service" here is Wayfinder's own engine, not a true
  external lookup.
- bulk-dataset-ingest's count outputs were written as plain int, but
  CalculationEvaluator.ValuesEqual only coerces "=" when BOTH sides are decimal —
  a boxed int silently compared unequal to a numeric literal. Fixed to decimal,
  matching ToFieldValues' own established convention for numbers in FieldValues.

Verified for real, not just unit tests: two live Playwright specs
(bulk-data-review-live.spec.ts alongside the existing support-systems-live.spec.ts)
against the actual Aspire-orchestrated stack both pass — a caseworker uploads a
real CSV, SafetyNet Underwriting's own separately-running app genuinely flags an
unrecognised tier, the client-fetched card UI shows it, the correction is saved
and resubmitted (materializing the corrected dataset, not the original upload),
and "Accept and finish" only becomes reachable once the real revalidation comes
back clean.

+++ Wayfinder.Engine/Services/ProcessManagerEngine.cs
@@ -2535,7 +2535,15 @@ public class ProcessManagerEngine : IProcessManager
         var fieldKey = action.Parameters[paramName]?.GetValue<string>();
         if (!string.IsNullOrWhiteSpace(fieldKey))
         {
-            updates[fieldKey] = value;
+            // decimal, not int: CalculationEvaluator.ValuesEqual only coerces when BOTH sides of
+            // "=" are decimal — a numeric literal in a showWhen/calculation expression parses as
+            // decimal, so a plain boxed int here would silently compare unequal to it even when
+            // numerically equal (found live: njf-contributions.json's "Accept and finish" route,
+            // showWhen: "contributionsErrorCount = 0", never became visible even once the count
+            // genuinely reached zero). Matches ToFieldValues' own JsonValue-decimal handling —
+            // "decimal for numbers in FieldValues" is this engine's established convention, not
+            // something to work around per call site.
+            updates[fieldKey] = (decimal)value;
         }
     }
